# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Chain": {
        "enum": [
          "bitcoin",
          "ethereum",
          "tron",
          "bnb",
          "polygon",
          "solana"
        ],
        "title": "Chain",
        "type": "string"
      },
      "EvidenceOut": {
        "additionalProperties": false,
        "properties": {
          "block_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Hash"
          },
          "block_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Height"
          },
          "captured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          }
        },
        "required": [
          "provider",
          "snapshot_id"
        ],
        "title": "EvidenceOut",
        "type": "object"
      },
      "GraphEdgeOut": {
        "additionalProperties": false,
        "properties": {
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "block_height": {
            "title": "Block Height",
            "type": "integer"
          },
          "from": {
            "title": "From",
            "type": "string"
          },
          "taint": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Taint",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "to": {
            "title": "To",
            "type": "string"
          },
          "tx_hash": {
            "title": "Tx Hash",
            "type": "string"
          },
          "value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "from",
          "to",
          "value",
          "asset",
          "taint",
          "tx_hash",
          "block_height",
          "timestamp"
        ],
        "title": "GraphEdgeOut",
        "type": "object"
      },
      "GraphNodeOut": {
        "additionalProperties": false,
        "properties": {
          "chain": {
            "$ref": "#/components/schemas/Chain"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "risk": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk"
          },
          "typologies": {
            "items": {
              "type": "string"
            },
            "title": "Typologies",
            "type": "array"
          },
          "vasp_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vasp Name"
          },
          "verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified"
          }
        },
        "required": [
          "id",
          "chain",
          "kind",
          "risk",
          "cluster_id",
          "vasp_name",
          "verified",
          "typologies"
        ],
        "title": "GraphNodeOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaintModel": {
        "enum": [
          "haircut",
          "poison"
        ],
        "title": "TaintModel",
        "type": "string"
      },
...
      "TraceAccepted": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TraceStatus"
          },
          "stream_url": {
            "title": "Stream Url",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "status",
          "stream_url"
        ],
        "title": "TraceAccepted",
        "type": "object"
      },
      "TraceGraphResponse": {
        "additionalProperties": false,
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeOut"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeOut"
            },
            "title": "Nodes",
            "type": "array"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "nodes",
          "edges"
        ],
        "title": "TraceGraphResponse",
        "type": "object"
      },
      "TraceParamsIn": {
        "additionalProperties": false,
        "properties": {
          "max_hops": {
            "default": 8,
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Max Hops",
            "type": "integer"
          },
          "min_taint": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "0.01",
            "title": "Min Taint"
          },
          "min_value": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "10",
            "title": "Min Value"
          },
          "taint_model": {
            "$ref": "#/components/schemas/TaintModel",
            "default": "haircut"
          }
        },
        "title": "TraceParamsIn",
        "type": "object"
      },
      "TraceRequest": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "minLength": 1,
            "title": "Address",
            "type": "string"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "chain": {
            "$ref": "#/components/schemas/Chain",
            "default": "tron"
          },
          "params": {
            "$ref": "#/components/schemas/TraceParamsIn"
          }
        },
        "required": [
          "address"
        ],
        "title": "TraceRequest",
        "type": "object"
      },
      "TraceResultOut": {
        "additionalProperties": false,
        "properties": {
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "trail_events": {
            "items": {
              "$ref": "#/components/schemas/TrailEventOut"
            },
            "title": "Trail Events",
            "type": "array"
          },
          "typologies": {
            "items": {
              "$ref": "#/components/schemas/TypologyOut"
            },
            "title": "Typologies",
            "type": "array"
          },
          "vasp_candidates": {
            "items": {
              "$ref": "#/components/schemas/VaspCandidateOut"
            },
            "title": "Vasp Candidates",
            "type": "array"
          }
        },
        "required": [
          "summary",
          "vasp_candidates",
          "typologies",
          "trail_events"
        ],
        "title": "TraceResultOut",
        "type": "object"
      },
      "TraceStatus": {
        "enum": [
          "queued",
          "running",
          "done",
          "partial",
          "failed"
        ],
        "title": "TraceStatus",
        "type": "string"
      },
      "TraceStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "block_heights": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Block Heights",
            "type": "object"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "chain": {
            "$ref": "#/components/schemas/Chain"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TraceResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Hash"
          },
          "start_address": {
            "title": "Start Address",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/TraceStatus"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "case_id",
          "start_address",
          "chain",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "block_heights",
          "result_hash",
          "result",
          "error"
        ],
        "title": "TraceStatusResponse",
        "type": "object"
      },
      "TrailEventOut": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "asset_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Symbol"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "reason",
          "address",
          "asset_symbol",
          "amount",
          "timestamp"
        ],
        "title": "TrailEventOut",
        "type": "object"
      },
      "TypologyOut": {
        "additionalProperties": false,
        "properties": {
          "addresses": {
            "items": {
              "type": "string"
            },
            "title": "Addresses",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Score",
            "type": "string"
          }
        },
        "required": [
          "name",
          "score",
          "model",
          "addresses"
        ],
        "title": "TypologyOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VaspCandidateOut": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence",
            "type": "string"
          },
          "confidence_terms": {
            "anyOf": [
              {
                "additionalProperties": {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Terms"
          },
          "confidence_weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Weights"
          },
          "deposit_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Address"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "title": "Evidence",
            "type": "array"
          },
          "hops_from_seed": {
            "title": "Hops From Seed",
            "type": "integer"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "reaching_paths": {
            "title": "Reaching Paths",
            "type": "integer"
          },
          "signals": {
            "items": {
              "type": "string"
            },
            "title": "Signals",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "rank",
          "tier",
          "verified",
          "source",
          "confidence",
          "name",
          "label",
          "deposit_address",
          "hops_from_seed",
          "reaching_paths",
          "signals",
          "confidence_terms",
          "confidence_weights",
          "evidence"
        ],
        "title": "VaspCandidateOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Real-time crypto fraud attribution (SIH26183)",
    "title": "Aegis API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trace": {
      "post": {
        "operationId": "start_trace_api_v1_trace_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Trace",
        "tags": [
          "trace"
        ]
      }
    },
    "/api/v1/trace/{trace_id}": {
      "get": {
        "operationId": "get_trace_api_v1_trace__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Trace",
        "tags": [
          "trace"
        ]
      }
    },
    "/api/v1/trace/{trace_id}/graph": {
      "get": {
        "operationId": "get_trace_graph_api_v1_trace__trace_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceGraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Trace Graph",
        "tags": [
          "trace"
        ]
      }
    },
...
  }
}